refactor: costPolicy -> costpolicy
