+++ src/main/java/subway/application/costpolicy/AgeCost.java
@@ -1,4 +1,4 @@
-package subway.application.costPolicy;
+package subway.application.costpolicy;
 
 import java.util.Arrays;
