+++ src/main/java/subway/application/costpolicy/DistanceCostPolicyChain.java
@@ -1,4 +1,4 @@
-package subway.application.costPolicy;
+package subway.application.costpolicy;
 
 import org.springframework.stereotype.Component;
 import subway.domain.Path;
